// src/generated/ModelControlsSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ModelRouteId } from "./ModelRouteId";
import type { ModelRouteSummary } from "./ModelRouteSummary";
import type { ReasoningEffort } from "./ReasoningEffort";

export type ModelControlsSnapshot = { routes: Array<ModelRouteSummary>, selectedRoute: ModelRouteId, efforts: Array<ReasoningEffort>, selectedEffort: ReasoningEffort, appliesAfterCurrentStep: boolean, };